+++ csfunctions/events/__init__.py
@@ -20,7 +20,9 @@ from .custom_operations import (
 from .dialog_data import DocumentReleasedDialogData, PartReleasedDialogData
 from .document_blocked import DocumentBlockedData, DocumentBlockedEvent
 from .document_create_check import DocumentCreateCheckData, DocumentCreateCheckEvent
+from .document_created import DocumentCreatedData, DocumentCreatedEvent
 from .document_field_calculation import DocumentFieldCalculationData, DocumentFieldCalculationEvent
+from .document_indexed import DocumentIndexedData, DocumentIndexedEvent
 from .document_modify_check import DocumentModifyCheckData, DocumentModifyCheckEvent
 from .document_release_check import DocumentReleaseCheckData, DocumentReleaseCheckEvent
 from .document_released import DocumentReleasedData, DocumentReleasedEvent
@@ -35,17 +37,23 @@ from .engineering_change_status_changed import EngineeringChangeStatusChangedDat
 from .field_value_calculation import FieldValueCalculationData, FieldValueCalculationEvent
 from .part_blocked import PartBlockedData, PartBlockedEvent
 from .part_create_check import PartCreateCheckData, PartCreateCheckEvent
+from .part_created import PartCreatedData, PartCreatedEvent
 from .part_field_calculation import PartFieldCalculationData, PartFieldCalculationEvent
+from .part_indexed import PartIndexedData, PartIndexedEvent
 from .part_modify_check import PartModifyCheckData, PartModifyCheckEvent
 from .part_release_check import PartReleaseCheckData, PartReleaseCheckEvent
 from .part_released import PartReleasedData, PartReleasedEvent
 from .workflow_task_trigger import WorkflowTaskTriggerEvent, WorkflowTaskTriggerEventData
 
 Event = Annotated[
-    DocumentBlockedEvent
+    DocumentCreatedEvent
+    | DocumentIndexedEvent
+    | DocumentBlockedEvent
     | DocumentReleasedEvent
     | DocumentReleaseCheckEvent
     | DocumentFieldCalculationEvent
+    | PartCreatedEvent
+    | PartIndexedEvent
     | PartBlockedEvent
     | PartReleasedEvent
     | PartReleaseCheckEvent
@@ -75,10 +83,14 @@ Event = Annotated[
     Field(discriminator="name"),
 ]
 EventData = (
-    DocumentBlockedData
+    DocumentCreatedData
+    | DocumentIndexedData
+    | DocumentBlockedData
     | DocumentReleasedData
     | DocumentReleaseCheckData
     | DocumentFieldCalculationData
+    | PartCreatedData
+    | PartIndexedData
     | PartBlockedData
     | PartReleasedData
     | PartReleaseCheckData
@@ -108,10 +120,14 @@ EventData = (
 )
 
 __all__ = [
+    "DocumentCreatedEvent",
+    "DocumentIndexedEvent",
     "DocumentBlockedEvent",
     "DocumentReleasedEvent",
     "DocumentReleaseCheckEvent",
     "DocumentFieldCalculationEvent",
+    "PartCreatedEvent",
+    "PartIndexedEvent",
     "PartBlockedEvent",
     "PartReleasedEvent",
     "PartReleaseCheckEvent",
@@ -132,10 +148,14 @@ __all__ = [
     "ChangeRequestStatusChangedEvent",
     "ChangeRequestStatusChangeCheckEvent",
     "WorkflowTaskTriggerEvent",
+    "DocumentCreatedData",
+    "DocumentIndexedData",
     "DocumentBlockedData",
     "DocumentReleasedData",
     "DocumentReleaseCheckData",
     "DocumentFieldCalculationData",
+    "PartCreatedData",
+    "PartIndexedData",
     "PartBlockedData",
     "PartReleasedData",
     "PartReleaseCheckData",